Revert "Implement `InputEvent`"
This reverts commit ee86f111168bb0583b05e8d5ca8de45b598cbba5.
