+++ examples/context_layering.rs
@@ -13,7 +13,7 @@ fn main() {
     App::new()
         .add_plugins((
             DefaultPlugins,
-            EnhancedInputPlugins,
+            EnhancedInputPlugin,
             PlayerBoxPlugin,
             GamePlugin,
         ))
