+++ tests/require_reset.rs
@@ -4,7 +4,7 @@ use bevy_enhanced_input::prelude::*;
 #[test]
 fn layering() {
     let mut app = App::new();
-    app.add_plugins((MinimalPlugins, InputPlugin, EnhancedInputPlugins))
+    app.add_plugins((MinimalPlugins, InputPlugin, EnhancedInputPlugin))
         .add_input_context::<First>()
         .add_input_context::<Second>();
 
@@ -58,7 +58,7 @@ fn layering() {
 #[test]
 fn switching() {
     let mut app = App::new();
-    app.add_plugins((MinimalPlugins, InputPlugin, EnhancedInputPlugins))
+    app.add_plugins((MinimalPlugins, InputPlugin, EnhancedInputPlugin))
         .add_input_context::<First>()
         .add_input_context::<Second>();
